+++ CHANGELOG.md
@@ -1,10 +1,19 @@
 # Changelog
 
-
 ### #149 Story CORE-600 | Add metrics to the throughput apps without inspr 
 - features:
   - added duration metrics for writes and reads
   - configured deployments to correctly integrate with Inspr's Prometheus
+---
+
+### #148 Story CORE-593 | [BETA] Python client SendRequest / [BETA] Python client HandleRoute
+- feature:
+  - added send new request function to python rest package. Now you can pass the method as a parameter.
+  - created send request method in python client
+  - created handle route method in python client
+  - created a full route demo example
+- fix:
+  - update lb sidecar to not crash when not finding the route env variables (added a return statement when an error occured)
 ---
 
 ### #147 Story CORE-591 | Implement benchmarking dapp out of Inspr Stack
@@ -17,7 +26,7 @@
 ---
 
 ### #145 Story CORE-585 | [BETA] Document the BETA-Routes feature
-- doc
+- doc:
   - described how to use the routes on Inspr (beta)
 ---